+++ app/components/FireBaseMessaging.js
@@ -38,40 +38,37 @@ class FireBaseMessaging extends Component {
    * @returns {void}
    */
   async componentDidMount() {
-      this.listeners = [
-        fNotifications.onNotification(this.onNotification),
-        fNotifications.onNotificationDisplayed(this.onNotificationDisplayed),
-        fNotifications.onNotificationOpened(this.onNotificationOpened),
-        fMessaging.onTokenRefresh(this.onTokenRefresh),
-        fMessaging.onMessage(this.onMessage),
-      ];
-      this.appState = 'active';
-      this.notificationsCount = 0;
+    this.listeners = [
+      fNotifications.onNotification(this.onNotification),
+      fNotifications.onNotificationDisplayed(this.onNotificationDisplayed),
+      fNotifications.onNotificationOpened(this.onNotificationOpened),
+      fMessaging.onTokenRefresh(this.onTokenRefresh),
+      fMessaging.onMessage(this.onMessage),
+    ];
+    this.appState = 'active';
+    this.notificationsCount = 0;
+    // AppState.addEventListener('change', this.handleAppStateChange);
 
-      AppState.addEventListener('change', this.handleAppStateChange);
-
-      if (isAndroid) {
-        this.initAndroidChannel();
-      }
-
-      if (isIOS) {
-        this.notificationsCount = await this.getDeliveredNotificationsCount();
-        firebase.messaging().ios.registerForRemoteNotifications();
-      }
-
+    if (isAndroid) {
+      this.initAndroidChannel();
+    }
+    if (isIOS) {
       this.notificationsCount = await this.getDeliveredNotificationsCount();
+      firebase.messaging().ios.registerForRemoteNotifications();
+    }
+    // this.notificationsCount = await this.getDeliveredNotificationsCount();
+    AppState.addEventListener('change', this.handleAppStateChange);
 
 
-      this.requestPermissions();
-      this.props.setFCMToken(await fMessaging.getToken());
+    this.requestPermissions();
+    this.props.setFCMToken(await fMessaging.getToken());
 
-      const event = await fNotifications.getInitialNotification();
+    const event = await fNotifications.getInitialNotification();
 
-      if (event) {
-        this.openActivityByEventId(event);
-        // if (isAndroid) NativeModules.DevSettings.reload();
-
-      }
+    if (event) {
+      this.openActivityByEventId(event);
+      // if (isAndroid) NativeModules.DevSettings.reload();
+    }
   }
 
   /**
@@ -173,7 +170,14 @@ class FireBaseMessaging extends Component {
       return activity;
     }
 
-    const event = applet.schedule.events.find(({ id }) => id === eventId);
+    let event = {};
+
+    Object.keys(applet.schedule.events).forEach(key => {
+      const e = applet.schedule.events[key];
+      if (e.id === eventId) {
+        event = e;
+      }
+    });
 
     if (!event) {
       return null;
@@ -232,7 +236,6 @@ class FireBaseMessaging extends Component {
       const eventId = _.get(notificationObj, 'notification._data.event_id', '');
       const appletId = _.get(notificationObj, 'notification._data.applet_id', '');
       const activityId = _.get(notificationObj, 'notification._data.activity_id', '');
-
       // Ignore the notification if some data is missing.
       if (!eventId || !appletId || !activityId) return;
 
@@ -246,7 +249,14 @@ class FireBaseMessaging extends Component {
       }
 
       let activity = applet.activities.find(({ id }) => id.endsWith(activityId));
-      const event = applet.schedule.events.find(({ id }) => id.endsWith(eventId));
+      let event = {};
+
+      Object.keys(applet.schedule.events).forEach(key => {
+        const e = applet.schedule.events[key];
+        if (e.id.endsWith(eventId)) {
+          event = e;
+        }
+      });
 
       if (activity) {
         return this.prepareAndOpenActivity(applet, activity, event);