+++ app/components/screen/Widget.js
@@ -19,25 +19,18 @@ import {
   TextEntry,
   TimeRange,
   VisualStimulusResponse,
+  RadioPrizes,
 } from '../../widgets';
 import TimePicker from '../../widgets/TimeRange/TimePicker';
 import { setSelected } from '../../state/responses/responses.actions';
 import { currentAppletSelector } from '../../state/app/app.selectors';
+import {
+  currentAppletTokenBalanceSelector,
+} from "../../state/responses/responses.selectors";
 
 // const TOKEN_LOGGER_SCHEMA = 'https://raw.githubusercontent.com/ChildMindInstitute/TokenLogger_applet/master/protocols/TokenLogger/TokenLogger_schema';
 
-const Widget = ({
-  screen,
-  answer,
-  onChange,
-  applet,
-  isCurrent,
-  isSelected,
-  setSelected,
-  onPress,
-  onRelease,
-  onContentError,
-}) => {
+const Widget = ({ screen, answer, onChange, applet, isCurrent, isSelected, setSelected, onPress, onRelease, onContentError, appletTokenBalance }) => {
   const valueType = R.path(['valueConstraints', 'valueType'], screen);
 
   if (screen.inputType === 'radio'
@@ -63,7 +56,7 @@ const Widget = ({
         onSelected={setSelected}
         value={answer}
         selected={isSelected}
-        token={valueType && valueType.includes('token')}
+        token={valueType && valueType.includes("token")}
       />
     );
   }
@@ -219,9 +212,24 @@ const Widget = ({
     );
   }
   // markdown items are rendered in ScreenDisplay
-  if (screen.inputType === 'markdown-message') {
+  if (screen.inputType === 'markdownMessage') {
     return null;
   }
+
+  if (screen.inputType === 'prize'
+    && R.path(['valueConstraints', 'itemList'], screen)) {
+    return (
+      <RadioPrizes
+        config={screen.valueConstraints}
+        onChange={onChange}
+        onSelected={setSelected}
+        value={answer}
+        selected={isSelected}
+        tokenBalance={appletTokenBalance.cumulativeToken}
+      />
+    );
+  }
+
   const [oneShot, setOneShot] = useState(false);
   useEffect(() => {
     if (onContentError && !oneShot) {
@@ -229,6 +237,7 @@ const Widget = ({
       onContentError();
     }
   });
+  
   return <WidgetError />;
 };
 
@@ -249,11 +258,13 @@ Widget.propTypes = {
   isSelected: PropTypes.bool.isRequired,
   onPress: PropTypes.func,
   onRelease: PropTypes.func,
+  appletTokenBalance: PropTypes.object.isRequired,
 };
 
 const mapStateToProps = state => ({
   applet: currentAppletSelector(state),
   isSelected: state.responses.isSelected,
+  appletTokenBalance: currentAppletTokenBalanceSelector(state),
 });
 
 const mapDispatchToProps = {