+++ app/state/media/media.selectors.js
@@ -1,3 +1,4 @@
 import * as R from 'ramda';
 
 export const mediaMapSelector = R.path(['media', 'mediaMap']);
+export const currentMediaSelector = R.path(['media', 'currentMedia']);