+++ app/widgets/MultiSelect.js
@@ -2,10 +2,11 @@ import React, { Component } from "react";
 import { View, Image } from "react-native";
 import PropTypes from "prop-types";
 import * as R from "ramda";
-import { ListItem, Text } from "native-base";
-import { CheckBox } from "react-native-elements";
+import { ListItem, Text, Icon } from 'native-base';
+import { CheckBox } from 'react-native-elements';
 import { getURL } from "../services/helper";
 import { colors } from "../themes/colors";
+import { TooltipBox } from './TooltipBox';
 
 export class MultiSelect extends Component {
   static isValid(value = [], { minValue = 1, maxValue = Infinity }) {
@@ -41,6 +42,15 @@ export class MultiSelect extends Component {
             onPress={() => this.onAnswer(token ? item.name.en : item.value)}
             key={index}
           >
+            <View style={{ width: '8%' }}>
+              {item.description ? (
+                <TooltipBox text={item.description}>
+                  <Icon type="FontAwesome" name="question-circle" style={{color: '#016fbe', fontSize: 24, marginHorizontal: 0}} />
+                </TooltipBox>
+              ) : (
+                <View />
+              )}
+            </View>
             <View style={{ width: "85%" }}>
               <View style={{ width: "100%", flexDirection: "row" }}>
                 {item.image ? (